# packages/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CariCash Nova API
  version: 0.2.0
  description: |
    CariCash Nova — Barbados-based mobile payments platform API.

    ## Key Concepts
    - **Jurisdiction**: Barbados. Base currency: BBD. NO FX — cross-currency postings rejected.
    - **Closed-loop wallets**: Customers, agents, and merchants hold balances internally.
    - **Double-entry ledger**: Every transaction posts balanced journal entries (sum DR = sum CR).
    - **Idempotency**: All money-moving endpoints require `idempotency_key`. Retries with same key+payload return cached result; same key+different payload returns 409 DUPLICATE_IDEMPOTENCY_CONFLICT.
    - **Maker-checker**: Reversals, manual adjustments, overdraft facilities, fee/commission changes require dual approval (maker ≠ checker).
    - **Posting via Durable Object**: All money movement routes through a serialized Durable Object per wallet-per-currency.
  contact:
    name: CariCash Engineering
    email: engineering@caricash.bb
  license:
    name: MIT
servers:
  - url: /
    description: Current host
  - url: http://localhost:8787
    description: Local development (wrangler dev)
...
security: []
paths:
...
  /auth/customer/login:
    post:
      operationId: customerLogin
      tags: [Auth]
      summary: Customer login
      description: Authenticate a customer using MSISDN and PIN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerLoginRequest'
            example:
              msisdn: '+12465551234'
              pin: '1234'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                token: '01HZ...'
                actor_id: '01HZ...'
                actor_type: CUSTOMER
                session_id: '01HZ...'
                correlation_id: '01HZ...'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/agent/login:
    post:
      operationId: agentLogin
      tags: [Auth]
      summary: Agent login
      description: Authenticate an agent using agent_code and PIN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentLoginRequest'
            example:
              agent_code: 'AGT001'
              pin: '5678'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/merchant/login:
    post:
      operationId: merchantLogin
      tags: [Auth]
      summary: Merchant login
      description: Authenticate a merchant using store_code and PIN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantLoginRequest'
            example:
              store_code: 'STORE001'
              pin: '9012'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/staff/login:
    post:
      operationId: staffLogin
      tags: [Auth]
      summary: Staff login
      description: Authenticate staff using staff_code and PIN. (OAuth stub — real IdP integration planned.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffLoginRequest'
            example:
              staff_code: 'STAFF001'
              pin: '3456'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ============================================================
  #  CUSTOMERS
  # ============================================================
...
  /agents:
    post:
      operationId: createAgent
      tags: [Agents]
      summary: Register a new agent
      description: Create an agent account with agent_code, name, MSISDN, PIN, and agent_type. Automatically creates WALLET + CASH_FLOAT accounts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
            example:
              agent_code: 'AGT002'
              name: 'Agent Smith'
              msisdn: '+12465558888'
              pin: '5678'
              agent_type: 'STANDARD'
              correlation_id: '01HZ...'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAgentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Agent with this code already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [Agents]
      summary: Get agent details
      description: Retrieve agent profile by ID. Returns 501 (stub).
      parameters:
        - $ref: '#/components/parameters/agentId'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorResponse'
        '501':
          description: Not yet implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    CustomerAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/customer/login
    AgentAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/agent/login
    MerchantAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/merchant/login
    StaffAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/staff/login (OAuth stub)
  parameters:
...
    agentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
      description: Agent actor ID (ULID)
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Validation failed'
            code: VALIDATION_ERROR
            correlation_id: '01HZ...'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Invalid credentials'
            correlation_id: '01HZ...'
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Rate limit exceeded'
            correlation_id: '01HZ...'
  schemas:
...
    ActorType:
      type: string
      enum: [CUSTOMER, AGENT, MERCHANT, STAFF]
    AgentTypeEnum:
      type: string
      enum: [STANDARD, AGGREGATOR]
...
    KycStatus:
      type: string
      enum: [NOT_STARTED, PENDING, APPROVED, REJECTED]
    ActorState:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED, CLOSED]
...
    ErrorCodeEnum:
      type: string
      enum:
        - UNAUTHORIZED
        - FORBIDDEN
        - VALIDATION_ERROR
        - INSUFFICIENT_FUNDS
        - CROSS_CURRENCY_NOT_ALLOWED
        - LIMIT_EXCEEDED
        - KYC_REQUIRED
        - BLOCKLISTED
        - DUPLICATE_IDEMPOTENCY_CONFLICT
        - IDEMPOTENCY_IN_PROGRESS
        - MAKER_CHECKER_REQUIRED
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
        - UNBALANCED_JOURNAL
        - JOURNAL_NOT_FOUND
        - ACCOUNT_NOT_FOUND
        - MAKER_CHECKER_VIOLATION
        - APPROVAL_NOT_FOUND
        - APPROVAL_ALREADY_DECIDED
        - INVALID_CREDENTIALS
        - ACCOUNT_LOCKED
        - RATE_LIMIT_EXCEEDED
        - SESSION_EXPIRED
        - ACTOR_NOT_FOUND
        - ACTOR_SUSPENDED
        - DUPLICATE_IDENTIFIER
        - MISSING_REQUIRED_FIELD
        - RECONCILIATION_MISMATCH
        - INTEGRITY_VIOLATION
        - REPAIR_FAILED

    # ---- Error response ----
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          $ref: '#/components/schemas/ErrorCodeEnum'
        name:
          type: string
          description: Error class name
        correlation_id:
          type: string
        details:
          type: object
          additionalProperties: true
        retryable:
          type: boolean
          default: false

    # ---- Auth requests ----
    CustomerLoginRequest:
      type: object
      required: [msisdn, pin]
      properties:
        msisdn:
          type: string
          description: Customer phone number
        pin:
          type: string
          description: Customer PIN
    AgentLoginRequest:
      type: object
      required: [agent_code, pin]
      properties:
        agent_code:
          type: string
          description: Agent identifier code
        pin:
          type: string
          description: Agent PIN
    MerchantLoginRequest:
      type: object
      required: [store_code, pin]
      properties:
        store_code:
          type: string
          description: Store identifier code
        pin:
          type: string
          description: Merchant PIN
    StaffLoginRequest:
      type: object
      required: [staff_code, pin]
      properties:
        staff_code:
          type: string
          description: Staff identifier code
        pin:
          type: string
          description: Staff PIN
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Bearer token for subsequent requests
        actor_id:
          type: string
        actor_type:
          $ref: '#/components/schemas/ActorType'
        session_id:
          type: string
        correlation_id:
          type: string

    # ---- Actor objects ----
    Actor:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/ActorType'
        state:
          $ref: '#/components/schemas/ActorState'
        msisdn:
          type: string
        agent_code:
          type: string
        store_code:
          type: string
        staff_code:
          type: string
        name:
          type: string
        first_name:
          type: string
        middle_name:
          type: string
        last_name:
          type: string
        display_name:
          type: string
        email:
          type: string
          format: email
        kyc_state:
          $ref: '#/components/schemas/KycStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ActorResponse:
      type: object
      properties:
        actor:
          $ref: '#/components/schemas/Actor'
        correlation_id:
          type: string

    # ---- Customer ----
...
    CreateAgentRequest:
      type: object
      required: [agent_code, name, msisdn, pin, agent_type]
      properties:
        agent_code:
          type: string
        name:
          type: string
        msisdn:
          type: string
        pin:
          type: string
        agent_type:
          $ref: '#/components/schemas/AgentTypeEnum'
        correlation_id:
          type: string
    CreateAgentResponse:
      type: object
      properties:
        actor:
          $ref: '#/components/schemas/Actor'
        wallet_id:
          type: string
        cash_float_id:
          type: string
        correlation_id:
          type: string

    # ---- Merchant ----
...